refactor(WidgetComposables): migrate utility functions to composables for widget navigation, formatting, and type categorization, enhancing code organization and maintainability
