+++ src/composables/__tests__/useWidgetNavigation.spec.js
