+++ src/composables/__tests__/useWidgetTypes.spec.js
