+++ src/composables/useWidgetFormatting.js
