+++ src/composables/useWidgetNavigation.js
