+++ src/composables/useWidgetTypes.js
